+++ src/models/mapinfo.ts
@@ -1,0 +1,5 @@
+export interface IMapInfo {
+    width: number;
+    height: number;
+    name: string;
+}